+++ tests/rdl_macro_test.rs
@@ -197,12 +197,13 @@ fn pipe_single_parent() {
   let w = fn_widget! {
     let edges = EdgeInsets::all(5.);
     let blank = pipe! {
-      move || -> Box<dyn SingleChild> {
-        if *$outside_blank {
+      fn_widget! {
+        let w: Box<dyn SingleChild>= if *$outside_blank {
           Box::new(Margin { margin: edges })
         } else {
           Box::new(FittedBox::new(BoxFit::None))
-        }
+        };
+        w
       }
     };
     rdl!{
@@ -229,12 +230,13 @@ fn pipe_multi_parent() {
   let stack_or_flex2 = stack_or_flex.clone_writer();
   let w = fn_widget! {
     let container = pipe! {
-      move || -> Box<dyn MultiChild> {
-        if *$stack_or_flex {
+      fn_widget!{
+        let w: Box<dyn MultiChild> = if *$stack_or_flex {
           Box::new(rdl!{ Stack { } })
         } else {
           Box::new(rdl!{ Flex { } })
-        }
+        };
+        w
       }
     };
 
@@ -264,7 +266,7 @@ fn pipe_as_child() {
   let w = fn_widget! {
     let blank = pipe!{
       $box_or_not2.then(|| {
-        move || {
+        fn_widget!{
           rdl!{ SizedBox { size: Size::new(100., 100.) } }
         }
       })
@@ -365,7 +367,7 @@ fn closure_in_fn_widget_capture() {
   let hi_res = Stateful::new(CowArc::borrowed(""));
   let hi_res2 = hi_res.clone_reader();
   let w = fn_widget! {
-    let text = @ Text { text: "hi" };
+    let mut text = @ Text { text: "hi" };
     let on_mounted = move |_: &mut _| *$hi_res.write() =$text.text.clone();
     @ $text { on_mounted }
   };
@@ -742,7 +744,7 @@ fn fix_subscribe_cancel_after_widget_drop() {
   let w = fn_widget! {
     let mut container = @SizedBox { size: Size::zero() };
     let h = watch!(*$trigger).subscribe(move |_| *$w_cnt.write() +=1 );
-    container = container.on_disposed(move |_| h.unsubscribe());
+    container.on_disposed(move |_| h.unsubscribe());
 
     @$container {
       @ {
@@ -848,7 +850,7 @@ fn fix_direct_use_map_writer_with_builtin() {
 #[test]
 fn fix_use_var_in_children() {
   let _w = fn_widget! {
-    let p = @MockBox { size: Size::zero() };
+    let mut p = @MockBox { size: Size::zero() };
     @ $p {
       opacity: 1.,
       // Use layout size query write of `p`