+++ themes/material/src/classes/buttons_cls.rs
@@ -25,15 +25,15 @@ named_style_impl!(common_label_only => {
 
 fn text_button_init(classes: &mut Classes) {
   fn interactive(w: Widget) -> Widget {
-    FatObj::new(base_interactive(w, md::RADIUS_20))
-      .foreground(BuildCtx::color())
-      .clamp(BTN_40_CLAMP)
-      .into_widget()
+    let mut w = FatObj::new(base_interactive(w, md::RADIUS_20));
+    w.foreground(BuildCtx::color())
+      .clamp(BTN_40_CLAMP);
+    w.into_widget()
   }
 
   classes.insert(
     TEXT_BTN,
-    multi_class![
+    class_multi_impl![
       style_class! { padding: md::EDGES_HOR_12, text_style: btn_label_style(18.) },
       interactive
     ],
@@ -44,58 +44,56 @@ fn text_button_init(classes: &mut Classes) {
 
   classes.insert(
     TEXT_BTN_LABEL_ONLY,
-    multi_class![
+    class_multi_impl![
       style_class! { padding: md::EDGES_HOR_12, text_style: btn_label_style(40.) },
       interactive
     ],
   );
-  classes.insert(TEXT_BTN_ICON_ONLY, multi_class![common_icon_only, interactive]);
+  classes.insert(TEXT_BTN_ICON_ONLY, class_multi_impl![common_icon_only, interactive]);
 }
 
 fn filled_button_init(classes: &mut Classes) {
   fn filled_interactive(w: Widget) -> Widget {
     let color = BuildCtx::color();
-    let w = FatObj::new(w)
-      .background(color)
+    let mut w = FatObj::new(w);
+    w.background(color)
       .radius(md::RADIUS_20)
-      .clamp(BTN_40_CLAMP)
-      .into_widget();
+      .clamp(BTN_40_CLAMP);
 
-    FatObj::new(base_interactive(w, md::RADIUS_20))
-      .foreground(BuildCtx::color().on_this_color(BuildCtx::get()))
-      .into_widget()
+    let mut w = FatObj::new(base_interactive(w.into_widget(), md::RADIUS_20));
+    w.foreground(BuildCtx::color().on_this_color(BuildCtx::get()));
+    w.into_widget()
   }
 
-  classes.insert(FILLED_BTN, multi_class![common_btn, filled_interactive]);
+  classes.insert(FILLED_BTN, class_multi_impl![common_btn, filled_interactive]);
   classes.insert(FILLED_BTN_LABEL, common_btn_label);
   classes.insert(FILLED_BTN_LEADING_ICON, empty_cls);
   classes.insert(FILLED_BTN_TRAILING_ICON, empty_cls);
 
-  classes.insert(FILLED_BTN_LABEL_ONLY, multi_class![common_label_only, filled_interactive]);
-  classes.insert(FILLED_BTN_ICON_ONLY, multi_class![common_icon_only, filled_interactive]);
+  classes.insert(FILLED_BTN_LABEL_ONLY, class_multi_impl![common_label_only, filled_interactive]);
+  classes.insert(FILLED_BTN_ICON_ONLY, class_multi_impl![common_icon_only, filled_interactive]);
 }
 
 fn button_init(classes: &mut Classes) {
   fn btn_interactive(w: Widget) -> Widget {
     let outline = Palette::of(BuildCtx::get()).outline();
-    let w = FatObj::new(w)
-      .border(Border::all(BorderSide { color: outline.into(), width: 1. }))
+    let mut w = FatObj::new(w);
+    w.border(Border::all(BorderSide { color: outline.into(), width: 1. }))
       .radius(md::RADIUS_20)
-      .clamp(BTN_40_CLAMP)
-      .into_widget();
+      .clamp(BTN_40_CLAMP);
 
-    FatObj::new(base_interactive(w, md::RADIUS_20))
-      .foreground(BuildCtx::color())
-      .into_widget()
+    let mut w = FatObj::new(base_interactive(w.into_widget(), md::RADIUS_20));
+    w.foreground(BuildCtx::color());
+    w.into_widget()
   }
 
-  classes.insert(BUTTON, multi_class![common_btn, btn_interactive]);
+  classes.insert(BUTTON, class_multi_impl![common_btn, btn_interactive]);
   classes.insert(BTN_LABEL, common_btn_label);
   classes.insert(BTN_LEADING_ICON, empty_cls);
   classes.insert(BTN_TRAILING_ICON, empty_cls);
 
-  classes.insert(BTN_LABEL_ONLY, multi_class![common_label_only, btn_interactive]);
-  classes.insert(BTN_ICON_ONLY, multi_class![common_icon_only, btn_interactive]);
+  classes.insert(BTN_LABEL_ONLY, class_multi_impl![common_label_only, btn_interactive]);
+  classes.insert(BTN_ICON_ONLY, class_multi_impl![common_icon_only, btn_interactive]);
 }
 
 fn fab_init(classes: &mut Classes) {
@@ -133,19 +131,19 @@ fn fab_init(classes: &mut Classes) {
       FabSize::Large => Radius::all(28.),
     };
 
-    let w = FatObj::new(w)
-      .background(background)
+    let mut w = FatObj::new(w);
+    w.background(background)
       .clamp(BoxClamp::min_width(btn_height).with_fixed_height(btn_height))
       .radius(radius);
 
-    FatObj::new(base_interactive(w.into_widget(), radius))
-      .foreground(foreground)
-      .into_widget()
+    let mut w = FatObj::new(base_interactive(w.into_widget(), radius));
+    w.foreground(foreground);
+    w.into_widget()
   }
 
   classes.insert(
     FAB_ICON_ONLY,
-    multi_class![
+    class_multi_impl![
       match fab_size() {
         FabSize::Large => style_class! { text_line_height: LARGE_ICON_SIZE },
         _ => common_icon_only,
@@ -156,7 +154,7 @@ fn fab_init(classes: &mut Classes) {
 
   classes.insert(
     FAB_LABEL_ONLY,
-    multi_class![
+    class_multi_impl![
       match fab_size() {
         FabSize::Mini => common_label_only,
         FabSize::Normal =>
@@ -172,7 +170,7 @@ fn fab_init(classes: &mut Classes) {
 
   classes.insert(
     FAB,
-    multi_class![
+    class_multi_impl![
       match fab_size() {
         FabSize::Mini => common_btn,
         FabSize::Normal => style_class! {
@@ -190,9 +188,11 @@ fn fab_init(classes: &mut Classes) {
   classes.insert(FAB_LEADING_ICON, empty_cls);
   classes.insert(FAB_TRAILING_ICON, empty_cls);
   classes.insert(FAB_LABEL, |w| match fab_size() {
-    FabSize::Large => FatObj::new(w)
-      .padding(md::EDGES_HOR_16)
-      .into_widget(),
+    FabSize::Large => {
+      let mut w = FatObj::new(w);
+      w.padding(md::EDGES_HOR_16);
+      w.into_widget()
+    }
     _ => common_btn_label(w),
   });
 }
@@ -206,9 +206,9 @@ fn btn_label_style(line_height: f32) -> TextStyle {
 
 fn base_interactive(w: Widget, radius: Radius) -> Widget {
   if DisabledRipple::get(BuildCtx::get()) {
-    FatObj::new(w)
-      .cursor(CursorIcon::Pointer)
-      .into_widget()
+    let mut w = FatObj::new(w);
+    w.cursor(CursorIcon::Pointer);
+    w.into_widget()
   } else {
     let hover_layer = HoverLayer::tracked(LayerArea::WidgetCover(radius));
     ripple! {